requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.4
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
kombu==5.3.7
openpyxl==3.1.2
phonenumbers==8.13.34
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
redis==5.0.3
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
